Restructure thor.orbits to make use of the simplified backend system
