+++ thor/orbits/__init__.py
@@ -2,3 +2,12 @@ from .kepler import *
 from .handler import *
 from .state import *
 from .orbits import *
+from .universal_propagate import *
+from .aberrations import *
+from .universal_ephemeris import *
+from .iterators import *
+from .gauss import *
+from .gibbs import *
+from .herrick_gibbs import *
+from .lambert import *
+from .iod import *
