+++ thor/orbits/aberrations.py
@@ -3,8 +3,8 @@ import numpy as np
 from numba import jit
 from numba.core.errors import NumbaPerformanceWarning
 
-from ...constants import Constants as c
-from ..propagate import propagateUniversal
+from ..constants import Constants as c
+from .universal_propagate import propagateUniversal
 
 # Numba will warn that numpy dot performs better on contiguous arrays. Fixing this warning
 # involves slicing numpy arrays along their second dimension which is unsupported
