--- thor/orbits/ephemeris/__init__.py
@@ -1,3 +1,0 @@
-from .aberrations import *
-from .universal import *
-from .ephemeris import *
