--- thor/orbits/iod/__init__.py
@@ -1,6 +1,0 @@
-from .iterators import *
-from .gauss import *
-from .gibbs import *
-from .herrick_gibbs import *
-from .lambert import *
-from .iod import *
