--- thor/orbits/propagate/__init__.py
@@ -1,3 +1,0 @@
-from .stumpff import *
-from .universal import *
-from .propagate import *
