+++ thor/orbits/tests/test_propagate.py
@@ -2,9 +2,9 @@ import numpy as np
 from astropy.time import Time
 from astropy import units as u
 
-from ....constants import Constants as c
-from ....utils import getHorizonsVectors
-from ....utils import testOrbits
+from ...constants import Constants as c
+from ...utils import getHorizonsVectors
+from ...testing import testOrbits
 from ..propagate import propagateOrbits
 
 MU = c.G * c.M_SUN
