+++ thor/orbits/universal_propagate.py
@@ -3,8 +3,11 @@ import numpy as np
 from numba import jit
 from numba.core.errors import NumbaPerformanceWarning
 
-from ...constants import Constants as c
+from ..constants import Constants as c
+from ..coordinates import transformCoordinates
+from ..utils import _checkTime
 from .stumpff import calcStumpff
+from .state import shiftOrbitsOrigin
 
 # Numba will warn that numpy dot performs better on contiguous arrays. Fixing this warning
 # involves slicing numpy arrays along their second dimension which is unsupported 
@@ -13,7 +16,7 @@ warnings.filterwarnings("ignore", category=NumbaPerformanceWarning)
 
 __all__ = [
     "calcChi",
-    "propagateUniversal"
+    "propagateUniversal",
 ]
 
 MU = c.G * c.M_SUN
